requirements/base.txt
Django==4.0.2
django-crispy-forms==1.14.0
django-import-export==2.7.1
django-axes==5.31.0
Pillow==9.3.0
crispy-bootstrap5==0.7
django-environ==0.9.0
psycopg2-binary==2.9.5
django-allauth==0.51.0
